~UPDATE: i18n locales to function
