+++ src/components/LocaleSwitcher.vue
@@ -1,38 +1,22 @@
 <template>
-<div class="locale-switcher">
-  <v-select
-      v-model=$i18n.locale
-      :items="languages()"
-      label="Select"
-  >
-  </v-select>
-</div>
+  <div class="locale-switcher">
+    <v-select v-model="$i18n.locale" :items="languages()" label="Select">
+    </v-select>
+  </div>
 </template>
 
 <script>
+import { locales } from "@/services/i18nconfiguration";
 export default {
   // sort locale file entries by keys: https://codeshack.io/json-sorter/
   name: "locale-switcher",
-  data () {
+  data() {
     return {
-      languages: function () {
-        const locales = require.context('/src/locales', true, /[A-Za-z0-9-_,\s]+\.json$/i)
-        const availableLocales = []
-        locales.keys().forEach(key => {
-          const matched = key.match(/([A-Za-z0-9-_]+)\./i)
-          if (matched && matched.length > 1) {
-            const locale = matched[1]
-            availableLocales.push(locale)
-          }
-        })
-        console.log(availableLocales)
-        return availableLocales
-      }
-    }
-  }
-}
+      languages: () => locales()
+    };
+  },
+};
 </script>
 
 <style scoped>
-
 </style>
